requirements.txt
Flask
Flask-Login
Werkzeug
groq
opencv-python
pytesseract
Pillow
imagehash
pandas
openpyxl
python-dateutil
python-dotenv